[Route] Increased the Precision of Cost Terms
For very large circuits, it has been found that the cost terms grow to
such a degree that adding the (considerably smaller) edge cost terms had
no effect. This can cause problems for algorithms that require the
accuracy and may yield lower quality of results.

Changed the cost terms from float to double.

+++ vpr/src/base/vpr_types.h
@@ -1744,8 +1744,8 @@ struct t_rr_node_route_inf {
     RREdgeId prev_edge;
 
     float acc_cost;
-    float path_cost;
-    float backward_path_cost;
+    double path_cost;
+    double backward_path_cost;
 
   public: //Accessors
     short occ() const { return occ_; }
